Practice 1 - For Each and If Statement
Find the smallest and biggest numbers in an array

Use For Each, Assign and If Statements to find the minimum and maximum number in an array of Int32 elements and print it.

Note: You can initialize the Array with {7, 5, 2, 4, 3, 9} or with your own values, as long as they are integers.